+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Library/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Library/releases)
 
-- Added: Deep table copy utility
+- Added: Character and realm utility function with separate return values